# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "使用gin+vue进行开发的博客平台",
        "title": "Gin-Vue-Blog Swagger API接口文档",
        "contact": {},
        "version": "v1.0.0"
    },
    "host": "127.0.0.1:8888",
    "basePath": "/",
    "paths": {
        "/allArticles": {
            "get": {
                "description": "获取所有文章",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "获取所有文章",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
...
        "/getAllMenus": {
            "get": {
                "description": "获取所有菜单",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "获取所有菜单",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-array_model_MenuItem"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/isvalid": {
            "get": {
                "description": "是否登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "是否登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "登录返回的Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_GetUserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用户名密码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "用户名密码登录",
                "parameters": [
                    {
                        "description": "用户名, 密码",
                        "name": "userInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UserLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/loginWithEmail": {
            "post": {
                "description": "邮箱登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "邮箱登录",
                "parameters": [
                    {
                        "description": "邮箱, 密码, 验证码",
                        "name": "LoginWithEmailRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.LoginWithEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/refresh": {
            "post": {
                "description": "刷新登录令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "刷新登录令牌",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户Refresh Token",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-model_TokenResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        },
        "/register": {
            "post": {
                "description": "注册普通用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "注册普通用户",
                "parameters": [
                    {
                        "description": "用户注册信息",
                        "name": "registerInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RegisterUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.CommonResponse-any"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "errhandle.ErrCode": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                4,
                5,
                6,
                7,
                8,
                9,
                10,
                11,
                12,
                13,
                14,
                15
            ],
            "x-enum-varnames": [
                "NoError",
                "InnerError",
                "TokenError",
                "ParamsError",
                "UserExists",
                "UserNonExists",
                "PasswordInvalid",
                "PermissionDenied",
                "SexError",
                "NameExists",
                "EmailFormatError",
                "PasswordTooShort",
                "FileSizeTooLarge",
                "FileTypeNoMatch",
                "VerifyCodeError",
                "OtherError"
            ]
        },
...
        "model.CommonResponse-any": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.CommonResponse-array_model_MenuItem": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MenuItem"
                    }
                },
                "msg": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.CommonResponse-model_GetUserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.GetUserResponse"
                },
                "msg": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.CommonResponse-model_TokenResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/model.TokenResponse"
                },
                "msg": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
        "model.CommonResponse-string": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/errhandle.ErrCode"
                }
            }
        },
...
        "model.GetUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
...
        "model.LoginWithEmailRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "model.MenuItem": {
            "type": "object",
            "properties": {
                "children": {
                    "description": "子菜单项列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.MenuItem"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "parent_id": {
                    "description": "父菜单项ID，用于构建多级菜单",
                    "type": "integer"
                },
                "sort": {
                    "description": "菜单的排序",
                    "type": "integer"
                },
                "title": {
                    "description": "菜单项标题",
                    "type": "string"
                },
                "url": {
                    "description": "菜单项链接",
                    "type": "string"
                }
            }
        },
...
        "model.RegisterUserRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nick_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                }
            }
        },
...
        "model.TokenResponse": {
            "type": "object",
            "properties": {
                "expiredAt": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "scope": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "model.UserLoginRequest": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}